+++ dev/dev.ino
@@ -5,14 +5,13 @@
 #include <FastLED.h>
 
 // How many leds in your strip?
-#define NUM_LEDS 1
+#define NUM_LEDS 10
 
 // For led chips like WS2812, which have a data line, ground, and power, you just
 // need to define DATA_PIN.  For led chipsets that are SPI based (four wires - data, clock,
 // ground, and power), like the LPD8806 define both DATA_PIN and CLOCK_PIN
 // Clock pin only needed for SPI based chipsets when not using hardware SPI
-#define DATA_PIN 3
-#define CLOCK_PIN 13
+#define DATA_PIN 9
 
 // Define the array of leds
 CRGB leds[NUM_LEDS];
@@ -22,12 +21,12 @@ void setup() {
 }
 
 void loop() { 
-  // Turn the LED on, then pause
-  leds[0] = CRGB::Red;
+  // Fill the LED array with red, then show
+  fill_solid(leds, NUM_LEDS, CRGB::Red);
   FastLED.show();
   delay(500);
-  // Now turn the LED off, then pause
-  leds[0] = CRGB::Black;
+  // Fill the LED array with black (off), then show
+  fill_solid(leds, NUM_LEDS, CRGB::Black);
   FastLED.show();
   delay(500);
 }
